Fix getULongAt bit pattern conversion on 32-bit platforms
Use Int64 instead of UInt to preserve full 64-bit bit pattern.
This fixes the CI failure in RuntimeNativeUnsignedByteArrayAccessorsTests.

+++ Sources/Runtime/RuntimeNativeAPI.swift
@@ -167,7 +167,7 @@ public func kk_native_byteArray_getULongAt(_ arrayRaw: Int, _ index: Int) -> Int
         byteCount: 8,
         functionName: "kk_native_byteArray_getULongAt"
     )
-    return Int(bitPattern: UInt(truncatingIfNeeded: value))
+    return Int(bitPattern: Int64(truncatingIfNeeded: value))
 }
 
 // MARK: - nativeHeap / nativeMemory allocation
